@@ -19,13 +19,14 @@
     ],
     "require": {
         "php": "^8.2",
-        "illuminate/support": "^11.0",
-        "symfony/polyfill-mbstring": "^1.29",
-        "spatie/laravel-html": "^3.6"
+        "illuminate/support": "^12.0",
+        "symfony/polyfill-mbstring": "^1.31",
+        "spatie/laravel-html": "^3.11"
     },
     "require-dev": {
-        "phpunit/phpunit": "^10.5",
-        "mockery/mockery": "^1.6"
+        "phpunit/phpunit": "^11.5.3",
+        "mockery/mockery": "^1.6",
+        "laravel/pint": "^1.21"
     },
     "autoload": {
         "psr-4": {
