@@ -9,9 +9,7 @@ use Illuminate\Support\Facades\Facade;
 class Autolink extends Facade
 {
     /**
-     * Get the registered name of the component.
-     *
-     * @return string
+     * {@inheritdoc}
      */
     protected static function getFacadeAccessor()
     {
