@@ -39,7 +39,7 @@ final class AutolinkTest extends TestCase
         return htmlentities($text, ENT_QUOTES, 'UTF-8', false);
     }
 
-    private function generateLink(string $title, string $href = null): string
+    private function generateLink(string $title, ?string $href = null): string
     {
         if (is_null($href)) {
             $href = $title;
@@ -52,12 +52,12 @@ final class AutolinkTest extends TestCase
         return '<a href="'.$this->e($href).'">'.$this->e($title).'</a>';
     }
 
-    public function testEscapesQuotes(): void
+    public function test_escapes_quotes(): void
     {
         $this->assertEquals('<a href="http://example.com/&quot;onmouseover=document.body.style.backgroundColor=&quot;pink&quot;;//">http://example.com/&quot;onmouseover=document.body.style.backgroundColor=&quot;pink&quot;;//</a>', $this->autolink->convert('http://example.com/"onmouseover=document.body.style.backgroundColor="pink";//'));
     }
 
-    public function testAutolinkWithSingleTrailingPunctuationAndSpace(): void
+    public function test_autolink_with_single_trailing_punctuation_and_space(): void
     {
         $url = 'http://example.com';
         $link = $this->generateLink('http://example.com');
@@ -69,14 +69,14 @@ final class AutolinkTest extends TestCase
         }
     }
 
-    public function testTerminatesOnAmpersand(): void
+    public function test_terminates_on_ampersand(): void
     {
         $url = 'http://example.com';
 
         $this->assertEquals("hello &#39;<a href=\"{$url}\">{$url}</a>&#39; hello", $this->autolink->convert("hello &#39;{$url}&#39; hello"));
     }
 
-    public function testAutolinkWithBrackets(): void
+    public function test_autolink_with_brackets(): void
     {
         $url1 = 'http://en.wikipedia.org/wiki/Sprite_(computer_graphics)';
         $link1 = $this->generateLink($url1);
@@ -97,7 +97,7 @@ final class AutolinkTest extends TestCase
         $this->assertEquals("{link: {$link3}}", $this->autolink->convert("{link: {$url3}}"));
     }
 
-    public function testAutolinkWithMultipleLinks(): void
+    public function test_autolink_with_multiple_links(): void
     {
         $url = 'http://example.com<br>http://example.com<br>http://example.com';
         $link = $this->generateLink($url);
@@ -105,7 +105,7 @@ final class AutolinkTest extends TestCase
         $this->assertEquals('<a href="http://example.com">http://example.com</a><br><a href="http://example.com">http://example.com</a><br><a href="http://example.com">http://example.com</a>', $this->autolink->convert($url));
     }
 
-    public function testAutolinkWithMultipleTrailingPunctuations(): void
+    public function test_autolink_with_multiple_trailing_punctuations(): void
     {
         $url = 'http://example.com';
         $link = $this->generateLink($url);
@@ -114,7 +114,7 @@ final class AutolinkTest extends TestCase
         $this->assertEquals("(link: {$link}).", $this->autolink->convert("(link: {$url})."));
     }
 
-    public function testAutolinkWithAlreadyLinked(): void
+    public function test_autolink_with_already_linked(): void
     {
         $link1 = $this->generateLink('Example', 'http://www.example.com');
         $link2 = '<a href="http://www.example.com">www.example.com</a>';
@@ -132,7 +132,7 @@ final class AutolinkTest extends TestCase
         $this->assertEquals($email, $this->autolink->convert($email));
     }
 
-    public function testAutolinkAtEol(): void
+    public function test_autolink_at_eol(): void
     {
         $url1 = 'http://example.com/foo.html';
         $url2 = 'http://example.com/bar.html';
@@ -140,7 +140,7 @@ final class AutolinkTest extends TestCase
         $this->assertEquals("<p><a href=\"{$url1}\">{$url1}</a><br /><a href=\"{$url2}\">{$url2}</a><br /></p>", $this->autolink->convert("<p>{$url1}<br />{$url2}<br /></p>"));
     }
 
-    public function testCallback(): void
+    public function test_callback(): void
     {
         $link = (string) $this->autolink->convert('Find ur favorite pokeman @ http://www.example.com', function ($element) {
             $this->assertInstanceOf(Element::class, $element);
@@ -156,7 +156,7 @@ final class AutolinkTest extends TestCase
         $this->assertEquals('Find ur favorite pokeman @ <a href="http://www.example.com">POKEMAN WEBSITE</a>', $link);
     }
 
-    public function testAutolinkWorks(): void
+    public function test_autolink_works(): void
     {
         $url = 'http://example.com/';
         $link = $this->generateLink($url);
@@ -164,12 +164,12 @@ final class AutolinkTest extends TestCase
         $this->assertEquals($link, $this->autolink->convert($url));
     }
 
-    public function testNotAutolinkWww(): void
+    public function test_not_autolink_www(): void
     {
         $this->assertEquals('Awww... man', $this->autolink->convert('Awww... man'));
     }
 
-    public function testDoesNotTerminateOnDash(): void
+    public function test_does_not_terminate_on_dash(): void
     {
         $url = 'http://example.com/Notification_Center-GitHub-20101108-140050.jpg';
         $link = $this->generateLink($url);
@@ -177,7 +177,7 @@ final class AutolinkTest extends TestCase
         $this->assertEquals($link, $this->autolink->convert($url));
     }
 
-    public function testDoesNotIncludeTrailingGt(): void
+    public function test_does_not_include_trailing_gt(): void
     {
         $url = 'http://example.com';
         $link = $this->generateLink($url);
@@ -185,7 +185,7 @@ final class AutolinkTest extends TestCase
         $this->assertEquals('&lt;'.$link.'&gt;', $this->autolink->convert('&lt;'.$url.'&gt;'));
     }
 
-    public function testLinksWithAnchors(): void
+    public function test_links_with_anchors(): void
     {
         $url = 'https://github.com/github/hubot/blob/master/scripts/cream.js#L20-20';
         $link = $this->generateLink($url);
@@ -193,7 +193,7 @@ final class AutolinkTest extends TestCase
         $this->assertEquals($link, $this->autolink->convert($url));
     }
 
-    public function testPolishWikipediaHaha(): void
+    public function test_polish_wikipedia_haha(): void
     {
         $url = 'https://pl.wikipedia.org/wiki/Komisja_śledcza_do_zbadania_sprawy_zarzutu_nielegalnego_wywierania_wpływu_na_funkcjonariuszy_policji,_służb_specjalnych,_prokuratorów_i_osoby_pełniące_funkcje_w_organach_wymiaru_sprawiedliwości';
 
@@ -204,7 +204,7 @@ final class AutolinkTest extends TestCase
         $this->assertEquals($expected, (string) $this->autolink->convert($input));
     }
 
-    public function testTheFamousNbsp(): void
+    public function test_the_famous_nbsp(): void
     {
         $url = 'http://google.com/';
 
@@ -215,7 +215,7 @@ final class AutolinkTest extends TestCase
         $this->assertEquals($expected, $this->autolink->convert($input));
     }
 
-    public function testDoesNotIncludeTrailingNonbreakingSpaces(): void
+    public function test_does_not_include_trailing_nonbreaking_spaces(): void
     {
         $url = 'http://example.com/';
         $link = $this->generateLink($url);
@@ -223,7 +223,7 @@ final class AutolinkTest extends TestCase
         $this->assertEquals("{$link}\xC2\xA0and", $this->autolink->convert("{$url}\xC2\xA0and"));
     }
 
-    public function testIdentifiesPreceedingNonbreakingSpaces(): void
+    public function test_identifies_preceeding_nonbreaking_spaces(): void
     {
         $url = 'http://example.com/';
         $link = $this->generateLink($url);
@@ -231,7 +231,7 @@ final class AutolinkTest extends TestCase
         $this->assertEquals("\xC2\xA0{$link} and", $this->autolink->convert("\xC2\xA0{$url} and"));
     }
 
-    public function testUrlsWith2WideUTF8Characters(): void
+    public function test_urls_with2_wide_ut_f8_characters(): void
     {
         $url = 'http://example.com/?foo=¥&bar=1';
         $link = $this->generateLink($url);
@@ -239,7 +239,7 @@ final class AutolinkTest extends TestCase
         $this->assertEquals("{$link} and", (string) $this->autolink->convert("{$url} and"));
     }
 
-    public function testUrlsWith4WideUTF8Characters(): void
+    public function test_urls_with4_wide_ut_f8_characters(): void
     {
         $url = 'http://example.com/?foo=&bar=1';
         $link = $this->generateLink($url);
@@ -247,7 +247,7 @@ final class AutolinkTest extends TestCase
         $this->assertEquals("{$link} and", $this->autolink->convert("{$url} and"));
     }
 
-    public function testHandlesUrlsWithEmojiProperly(): void
+    public function test_handles_urls_with_emoji_properly(): void
     {
         $url = 'http://foo.com/💖a';
         $link = $this->generateLink($url);
@@ -255,7 +255,7 @@ final class AutolinkTest extends TestCase
         $this->assertEquals("{$link} and", $this->autolink->convert("{$url} and"));
     }
 
-    public function testIdentifiesNonbreakingSpacesPreceedingEmails(): void
+    public function test_identifies_nonbreaking_spaces_preceeding_emails(): void
     {
         $email = 'foo@example.com';
         $link = $this->generateLink($email, "mailto:{$email}");
@@ -263,7 +263,7 @@ final class AutolinkTest extends TestCase
         $this->assertEquals("email\xC2\xA0{$link}", $this->autolink->convert("email\xC2\xA0{$email}"));
     }
 
-    public function testIdentifiesUnicodeSpaces(): void
+    public function test_identifies_unicode_spaces(): void
     {
         $this->assertEquals(
             "This is just a test. <a href=\"http://www.example.com\">http://www.example.com</a>\u{202F}\u{2028}\u{2001}",
@@ -271,7 +271,7 @@ final class AutolinkTest extends TestCase
         );
     }
 
-    public function testWwwIsCaseInsensitive(): void
+    public function test_www_is_case_insensitive(): void
     {
         $url = 'www.reddit.com';
         $link = $this->generateLink($url);
@@ -294,7 +294,7 @@ final class AutolinkTest extends TestCase
         $this->assertEquals($link, $this->autolink->convert($url));
     }
 
-    public function testNonEmailsEndingInPeriods(): void
+    public function test_non_emails_ending_in_periods(): void
     {
         $this->assertEquals('abc/def@ghi.', $this->autolink->convert('abc/def@ghi.'));
         $this->assertEquals('abc/def@ghi. ', $this->autolink->convert('abc/def@ghi. '));
@@ -304,7 +304,7 @@ final class AutolinkTest extends TestCase
         $this->assertEquals('abc/<a href="mailto:def@ghi.x">def@ghi.x</a>. a', $this->autolink->convert('abc/def@ghi.x. a'));
     }
 
-    public function testUrlsWithEntitiesAndParens(): void
+    public function test_urls_with_entities_and_parens(): void
     {
         $this->assertEquals('&lt;<a href="http://www.google.com">http://www.google.com</a>&gt;', $this->autolink->convert('&lt;http://www.google.com&gt;'));
         $this->assertEquals('&lt;<a href="http://www.google.com">http://www.google.com</a>&gt;)', $this->autolink->convert('&lt;http://www.google.com&gt;)'));
@@ -328,7 +328,7 @@ final class AutolinkTest extends TestCase
         $this->assertEquals("(URL is {$link}.)", $this->autolink->convert("(URL is {$url}.)"));
     }
 
-    public function testUrlsWithParens(): void
+    public function test_urls_with_parens(): void
     {
         $this->assertEquals('(<a href="http://example.com">http://example.com</a>)', $this->autolink->convert('(http://example.com)'));
         $this->assertEquals('((<a href="http://example.com/()">http://example.com/()</a>))', $this->autolink->convert('((http://example.com/()))'));
@@ -342,7 +342,7 @@ final class AutolinkTest extends TestCase
         $this->assertEquals('〈<a href="http://example.com/">http://example.com/</a>〉', $this->autolink->convert('〈http://example.com/〉'));
     }
 
-    public function testUrlsWithQuotes(): void
+    public function test_urls_with_quotes(): void
     {
         $this->assertEquals("'<a href=\"http://example.com\">http://example.com</a>'", $this->autolink->convert("'http://example.com'"));
         $this->assertEquals('"<a href="http://example.com">http://example.com</a>"', $this->autolink->convert('"http://example.com"'));
