@@ -11,7 +11,7 @@ use OsiemSiedem\Tests\Autolink\TestCase;
 
 final class WwwParserTest extends TestCase
 {
-    public function testParse(): void
+    public function test_parse(): void
     {
         $cursor = new Cursor('www.localhost');
         $parser = new WwwParser;
@@ -19,7 +19,7 @@ final class WwwParserTest extends TestCase
         $this->assertNull($parser->parse($cursor));
     }
 
-    public function testStopsParsingAtWhitespace(): void
+    public function test_stops_parsing_at_whitespace(): void
     {
         $cursor = new Cursor("www.example.com\n");
         $parser = new WwwParser;
@@ -27,7 +27,7 @@ final class WwwParserTest extends TestCase
         $this->assertInstanceOf(Element::class, $parser->parse($cursor));
     }
 
-    public function testMinimumLength(): void
+    public function test_minimum_length(): void
     {
         $cursor = new Cursor('www.a');
         $parser = new WwwParser;
